youtube_tutor/requirements.txt
streamlit==1.32.0
langchain==0.1.16
langchain-community==0.0.34
groq==0.5.0
youtube-transcript-api==0.6.2
chromadb==0.4.24
scikit-learn==1.4.2
numpy==1.26.4
python-dotenv==1.0.1
tiktoken==0.6.0
google-api-python-client==2.126.0
sentence-transformers==2.7.0